requirements.txt
numpy
pandas
requests
firebase_admin
flask
seaborn
matplotlib
plotly
colorama
pandas-alive 
attrs
ffmpeg
openpyxl
ipywidgets
latex
strictly_typed_pandas
click
tqdm
pyhumps
scikit-learn
greytheory
statsmodels
darts
tensorflow
nbformat>=4.2.0
stringcase